third_party/blink/web_tests/accessibility/aria-hidden-hides-all-elements.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
</head>
<body id="body">

<div id="content" role="group">

<main id="main" aria-hidden="true">
    <video controls="controls"></video>
    <select id="select"><option>a</select>
    <input type="range">
    <input type="data">
    <progress></progress>
    <ul id="ul"><li>item</li></ul>
    <select size="10"><option>a</select>
    <img src="resources/cake.png" width="145" height="126" usemap="#map">
    <map name="map">
        <area shape="rect" coords="0,0,100,100" href="#" alt="cake">
    </map>
</main>

</div>

<script>
test(function(t) {
    var content = accessibilityController.accessibleElementById("main");
    assert_equals(content.childrenCount, 8);
    assert_true(content.isIgnored);
    assert_false(content.childAtIndex(0).isIgnored);
    assert_false(content.childAtIndex(1).isIgnored);
    assert_false(content.childAtIndex(2).isIgnored);
    assert_false(content.childAtIndex(3).isIgnored);
    assert_true(content.childAtIndex(4).isIgnored);
    assert_true(content.childAtIndex(5).isIgnored);
    assert_false(content.childAtIndex(6).isIgnored);
    assert_false(content.childAtIndex(7).isIgnored);

    document.getElementById("ul").tabIndex = -1;

    assert_equals(content.childrenCount, 8,
                      "Making list focusable should not impact if it's hidden");
    assert_true(content.isIgnored);
    assert_false(content.childAtIndex(0).isIgnored);
    assert_false(content.childAtIndex(1).isIgnored);
    assert_false(content.childAtIndex(2).isIgnored);
    assert_false(content.childAtIndex(3).isIgnored);
    assert_true(content.childAtIndex(4).isIgnored);
    assert_false(content.childAtIndex(5).isIgnored);
    assert_false(content.childAtIndex(6).isIgnored);
    assert_false(content.childAtIndex(7).isIgnored);

    document.getElementById("ul").removeAttribute("tabindex");

    assert_equals(content.childrenCount, 8,
                          "Making list unfocusable should make it hidden again");
    assert_true(content.isIgnored);
    assert_false(content.childAtIndex(0).isIgnored);
    assert_false(content.childAtIndex(1).isIgnored);
    assert_false(content.childAtIndex(2).isIgnored);
    assert_false(content.childAtIndex(3).isIgnored);
    assert_true(content.childAtIndex(4).isIgnored);
    assert_true(content.childAtIndex(5).isIgnored);
    assert_false(content.childAtIndex(6).isIgnored);
    assert_false(content.childAtIndex(7).isIgnored);
});
</script>

</body>
</html>
